requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0
feedparser>=6.0.0

# LLM Integration
google-genai>=1.0.0

# Data Management
pandas>=2.1.0
openpyxl>=3.1.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
